fixed environment, factored out waterline manager
