+++ microservice/unit_tests/testorderdo.js
@@ -82,7 +82,7 @@ let test_create = {
     totalPrice: 500
 };
 
-//testA();
+testA();
 //testB();
 //testC(test_create);
 //testD()
